Move countWords to utils file
We are not using Program.py for now and this file uses en_core_web_sm which is not installed.
